refactor(mobile): extract activity detail screen hook
